src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/dynamically-created-shadow-root-expected.txt

CONSOLE ERROR: line 11: Uncaught TypeError: host1.createShadowRoot is not a function
